Bug 1786556 - [PerfCompare] update API endpoint to allow selecting a single revision
